fix: add service code , fix code appointment and calendar
