+++ src/app/interface/LocationVisit.ts
@@ -1,0 +1,5 @@
+export enum LocationVisit {
+    Online = 0,
+    InClinic = 1,
+    other = 2
+  }
