+++ src/app/interface/TypeVisit.ts
@@ -1,0 +1,6 @@
+export enum TypeVisit {
+    Firstconsultation,
+    FollowUp,
+    Emergency,
+    other
+  }
